requirements.txt
asgiref==3.4.1
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.5
cryptography==2.9.2
defusedxml==0.6.0
Django==3.2.6
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
